Solved problem 23
This one was pretty tricky. First I solved a bug in my old misc/factors
function which was missing out the square root of square numbers (this
apparently still got the right answer to 21), then I tried doing lots
of different stuff to get all of the abundant pairs quickly, most of my
attempts ending in far too slow programs. The solution (even though it
is neither very clever nor very pretty) was to use my trusted boolean
vector to cross the sums off, and using a loop-recur with lots of lets
to prevent unnecessary computation. It doesn't look too nice but it
actually gets the right answer in under 3s.

I also added the sum-bitmap function to the misc namespace.